requirements/base.txt
Django==3.2.5
python-decouple==3.4
django-environ==0.4.5
psycopg2==2.9.1
django-model-utils==4.1.1
django-cleanup==5.2.0
django-import-export==2.5.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
dj-rest-auth==2.1.8
djangorestframework-simplejwt==4.7.2
drf-writable-nested==0.6.3
djangorestframework-filters==1.0.0.dev0
drf-api-tracking==1.8.0
drf-renderer-xlsx==0.4.0
django-allauth==0.45.0
pdfkit==0.6.1
